fix : room 생성시 기존에 방있으면 지우고 만들어줌

See merge request s05-final/S05P31B304!229

+++ back_springboot/src/main/java/com/codeback/service/room/RoomService.java
@@ -53,6 +53,8 @@ public class RoomService {
 
     @Transactional
     public void deleteByUserNumber(long userNumber) {
-        roomRepository.deleteByUserUserNumber(userNumber);
+        Optional<Room> room = roomRepository.findByUserUserNumber(userNumber);
+        if(room.isPresent())
+            roomRepository.deleteByUserUserNumber(userNumber);
     }
 }
